Fix os join syntax (#6)
* Fix os join syntax

* test upgrade-charm with unit test

* don't search `cni_conf_dir` if it doesn't exist

+++ reactive/kubernetes_node_base.py
@@ -25,12 +25,14 @@ def upgrade_charm():
 
     # rename default CNI config to 01-default.*
     cni_conf_dir = "/etc/cni/net.d/"
+    if not os.path.exists(cni_conf_dir):
+        return
     for filename in os.listdir(cni_conf_dir):
         if filename.startswith("05-default."):
             new_filename = "01" + filename[2:]
             os.replace(
-                os.join.path(cni_conf_dir, filename),
-                os.join.path(cni_conf_dir, new_filename),
+                os.path.join(cni_conf_dir, filename),
+                os.path.join(cni_conf_dir, new_filename),
             )
